fix 7c

+++ pi_escape/es/component_structs.h
@@ -109,8 +109,8 @@ typedef struct ContainerComponent {
 } ContainerComponent;
 
 typedef struct ActivationComponent {
-	int currenttime;
-	int getto;
+	double currenttime;
+	double getto;
 } ActivationComponent;
 
 typedef struct ActivatableComponent {
